fixed a bug, wrong data type for cf_tol in gmres

+++ krylov/krylov.h
@@ -492,7 +492,7 @@ typedef struct
    int      stop_crit;
    int      converged;
    double   tol;
-   int      cf_tol;
+   double   cf_tol;
    double   rel_residual_norm;
 
    void  *A;
